eatgo-api/src/test/java/com/soobinnn/eatgo/ex03_eatgo/interfaces/RestaurantControllerTest.java
package com.soobinnn.eatgo.ex03_eatgo.interfaces;

import static org.hamcrest.CoreMatchers.containsString;
import static org.mockito.BDDMockito.given;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import com.soobinnn.eatgo.ex03_eatgo.application.RestaurantService;
import com.soobinnn.eatgo.ex03_eatgo.domain.MenuItem;
import com.soobinnn.eatgo.ex03_eatgo.domain.Restaurant;
import java.util.ArrayList;
import java.util.List;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

@RunWith(SpringRunner.class)
@WebMvcTest(RestaurantController.class)
public class RestaurantControllerTest {

  @Autowired private MockMvc mvc;

  @MockBean private RestaurantService restaurantService;
  //    @SpyBean(RestaurantRepositoryImpl.class)
  //    private RestaurantRepository restaurantRepository;
  //
  //    @SpyBean(MenuItemRepositoryImpl.class)
  //    private MenuItemRepository menuItemRepository;
  //
  //    @SpyBean(RestaurantService.class)
  //    private RestaurantService restaurantService;

  // perform은 예외가 나올 수 있으니 Exception을 해야함.
  // Controller는 이 Service를 활용한다 라는 이 사실에만 집중할 뿐이지
  // 이 Service가 어떻게 작동하는지는 중요하지 않다. - Mock Object를 사용하는 이유

  @Test
  public void list() throws Exception {
    List<Restaurant> restaurants = new ArrayList<>();
    restaurants.add(new Restaurant(1004L, "Joker House", "Seoul"));

    given(restaurantService.getRestaurants()).willReturn(restaurants);

    mvc.perform((get("/restaurants")))
        .andExpect(status().isOk())
        .andExpect(content().string(containsString("\"id\":1004")))
        .andExpect(content().string(containsString("\"name\":\"Joker House\"")));
  }

  @Test
  public void detail() throws Exception {
    Restaurant restaurant1 = new Restaurant(1004L, "JOKER House", "Seoul");
    restaurant1.addMenuItem(new MenuItem("Kimchi"));

    Restaurant restaurant2 = new Restaurant(2020L, "Cyber Food", "Seoul");

    given(restaurantService.getRestaurant(1004L)).willReturn(restaurant1);
    given(restaurantService.getRestaurant(2020L)).willReturn(restaurant2);

    mvc.perform(get("/restaurants/1004"))
        .andExpect(status().isOk())
        .andExpect(content().string(containsString("\"id\":1004")))
        .andExpect(content().string(containsString("\"name\":\"JOKER House\"")))
        .andExpect(content().string(containsString("Kimchi")));

    mvc.perform(get("/restaurants/2020"))
        .andExpect(status().isOk())
        .andExpect(content().string(containsString("\"id\":2020")))
        .andExpect(content().string(containsString("\"name\":\"Cyber Food\"")));
  }
}
